requirements.txt
paddlepaddle         #if use CPU
#paddlepaddle-gpu     #if use GPU
Pillow==9.5.0
opencv-python
pandas
numpy
shapely
pyclipper
pyyaml
flask
fitz
frontend
PyMuPDF==1.19
openpyxl
premailer
scikit-image
python-docx 
BeautifulSoup4
tools